feat: auto-delete in voice-chat feature (#72)
* added new config file to module config

* fixed text description

* added example config for voice-channel feature

* added in-voice chat purgeOnStart functionality;

* added filtering of channels (no duplicates in config); Added overwriting of voice-channels config

* added reusing of botReady unique channels to reduce work load

* [auto-delete] added new event and feature functionality in-voice chats

* [auto-delete] added "could-not-fetch-messages" locale

* [auto-delete] fixed missing import of localize function

* [auto-delete] deleted unnecessary object cloning

* [auto-delete] added catch functionality on channel.fetch

* [auto-delete] fixed if statement to be more readable

* [auto-delete] added channel message catch functionality

* [auto-delete] added early exit on no messages in channel

* [auto-delete] added now using variable for message parameter

* [auto-delete] fixed unnecessary new line

* [auto-delete] added channel message catch functionality + early return

* [auto-delete] added usage of channelMessage variable in bulkDelete

* [auto-delete] fixed if statement to be more readable

* [auto-delete] added channel.messages catch functionality + early return

* [auto-delete] added usage of channelMessage variable in Bulkdelete

* [auto-delete] fixed order of if statements to fit logically

* fixed eslint errors

* [auto-delete] fixed logic errors and restructured code

* fix(auto-delete): only parsing not int valus as int

* fix(auto-delete): changed early return statement

* fix(auto-delete): added check for Voice channel Type

* feat(auto-delete:config): added more detailed bilingual description and names

* fix(auto-delete): changed from "Zeit" to Timeout

* feat(auto-delete): added locale entries for could-not-fetch-messages

* fix(auto-delete): translated text to locale language

+++ modules/auto-delete/events/messageCreate.js
@@ -3,9 +3,9 @@ module.exports.run = async function (client, msg) {
     if (!msg.guild) return;
     if (msg.guild.id !== client.guildID) return;
     if (!msg.member) return;
+    if (!client.modules['auto-delete'].uniqueChannels) return;
 
-    const channels = client.configurations['auto-delete']['channels'];
-    const channel = channels.find(c => c.channelID === msg.channel.id);
+    const channel = client.modules['auto-delete'].uniqueChannels.find(c => c.channelID === msg.channel.id);
     if (!channel) return;
     setTimeout(() => {
         if (msg.deletable && !msg.pinned) msg.delete().catch(() => {
